pyproject.toml
[project]
name = "ha-lumagen"
version = "0.1.0"
requires-python = ">=3.13"

[dependency-groups]
dev = [
    "homeassistant",
    "pytest",
    "pre-commit",
    "textual>=1.0",
    "ty",
    "typer-slim",
    "ruamel-yaml>=0.19.1",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins — shadowing builtins
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "PL",     # pylint (ruff reimplementation)
    "RUF",    # ruff-specific
    "PIE",    # flake8-pie
    "RET",    # flake8-return
    "PERF",   # perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "T20",    # flake8-print — catch stray prints
    "PTH",    # flake8-use-pathlib
    "TID",    # flake8-tidy-imports
]

ignore = [
    "PLR09",  # too-many-* (arguments, branches, statements) — not useful for HA entities
    "PLR2004", # magic-value-comparison — too noisy for protocol code with numeric constants
    "PLW2901", # redefined-loop-variable — common pattern in asyncio
    "SIM108",  # use-ternary-expression — often less readable
]

[tool.ty.environment]
extra-paths = ["custom_components", "custom_components/lumagen"]

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["custom_components", "custom_components/lumagen"]
# HA's CachedProperties metaclass + multi-inheritance (CoordinatorEntity + platform
# entity) causes false positives on `available`, and narrowing `entity_description`
# to a subclass type is the standard HA pattern. Pyright can't reason through either.
reportIncompatibleVariableOverride = false

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = [
    "ignore:Inheritance class HomeAssistantApplication from web.Application is discouraged:DeprecationWarning",
]
